+++ lib/maze/hooks/command_hooks.rb
@@ -1,0 +1,8 @@
+# Contains logic for the Cucumber hooks when in Command mode
+module Maze
+  module Hooks
+    class CommandHooks < InternalHooks
+      # Nothing required at present
+    end
+  end
+end