+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# TBD
+# 5.8.0 - 2021/07/26
 
 ## Enhancements
 
-- Add a generic send keys method so you can send keys without finding an element [#278](https://github.com/bugsnag/maze-runner/pull/278)
+- Add Appium driver no-element send_keys method [#278](https://github.com/bugsnag/maze-runner/pull/278)
+- Add support for running tests against local browsers [#279](https://github.com/bugsnag/maze-runner/pull/279)
 
 # 5.7.0 - 2021/07/14